Add THHalfTensor support to cutorch.
