+++ CMakeLists.txt
@@ -11,7 +11,8 @@ IF (NOT WIN32)
 SET(CMAKE_C_FLAGS "-std=c99 -Werror=implicit-function-declaration ${CMAKE_C_FLAGS}")
 ENDIF (NOT WIN32)
 IF(CUDA_HAS_FP16 OR NOT ${CUDA_VERSION} LESS 7.5)
-  SET(CMAKE_C_FLAGS "-DCUDA_HAS_FP16=1 ${CMAKE_C_FLAGS}")
+  ADD_DEFINITIONS(-DTH_GENERIC_USE_HALF=1)
+  ADD_DEFINITIONS(-DCUDA_HAS_FP16=1)
 ENDIF()
 
 INCLUDE_DIRECTORIES(${CUDA_INCLUDE_DIRS})
