+++ init.lua
@@ -16,8 +16,8 @@ torch.CudaTensor.__tostring__        = torch.FloatTensor.__tostring__
 torch.CudaDoubleStorage.__tostring__ = torch.DoubleStorage.__tostring__
 torch.CudaDoubleTensor.__tostring__  = torch.DoubleTensor.__tostring__
 if cutorch.hasHalf then
-   torch.CudaHalfStorage.__tostring__  = torch.FloatStorage.__tostring__
-   torch.CudaHalfTensor.__tostring__  = torch.FloatTensor.__tostring__
+   torch.CudaHalfStorage.__tostring__  = torch.HalfStorage.__tostring__
+   torch.CudaHalfTensor.__tostring__  = torch.HalfTensor.__tostring__
 end
 
 require('cutorch.Tensor')
@@ -56,6 +56,20 @@ function cutorch.createCudaHostTensor(...)
    local storage = torch.FloatStorage(cutorch.CudaHostAllocator, size:prod())
    return torch.FloatTensor(storage, 1, size:storage())
 end
+
+function cutorch.createCudaHostDoubleTensor(...)
+   local size = longTensorSize(...)
+   local storage = torch.DoubleStorage(cutorch.CudaHostAllocator, size:prod())
+   return torch.DoubleTensor(storage, 1, size:storage())
+end
+
+if cutorch.hasHalf then
+  function cutorch.createCudaHostHalfTensor(...)
+     local size = longTensorSize(...)
+     local storage = torch.HalfStorage(cutorch.CudaHostAllocator, size:prod())
+     return torch.HalfTensor(storage, 1, size:storage())
+   end
+ end
 
 -- Creates a CudaTensor using the CudaUVAAllocator.
 -- Accepts either a LongStorage or a sequence of numbers.
