+++ lib/THC/generic/THCTensorCopy.c
@@ -4,7 +4,6 @@
 
 /* specific methods */
 
-#ifndef THC_REAL_IS_HALF
 void THCTensor_(copyCPU)(THCState *state, THCTensor *self, struct THTensor *src)
 {
   THArgCheck(THCTensor_(nElement)(state, self) == THTensor_(nElement)(src), 2, "sizes do not match");
@@ -22,9 +21,7 @@ void THCTensor_(copyCPU)(THCState *state, THCTensor *self, struct THTensor *src)
     THCTensor_(freeCopyTo)(state, selfc, self);
   }
 }
-#endif
 
-#ifndef THC_REAL_IS_HALF
 #define IMPLEMENT_TH_CUDA_TENSOR_COPY(TYPEC)                            \
 void THCTensor_(copy##TYPEC)(THCState *state, THCTensor *self, struct TH##TYPEC##Tensor *src)                \
 {                                                                       \
@@ -42,19 +39,6 @@ void THCTensor_(copy##TYPEC)(THCState *state, THCTensor *self, struct TH##TYPEC#
     THTensor_(free)(srcf);                                              \
   }                                                                     \
 }
-#else
-#define IMPLEMENT_TH_CUDA_TENSOR_COPY(TYPEC)                            \
-void THCTensor_(copy##TYPEC)(THCState *state, THCTensor *self, struct TH##TYPEC##Tensor *src)                \
-{                                                                       \
-  THArgCheck(THCTensor_(nElement)(state, self) == TH##TYPEC##Tensor_nElement(src), 2, "sizes do not match"); \
-  THLongStorage *size = TH##TYPEC##Tensor_newSizeOf(src);               \
-  THCudaTensor *buffer = THCudaTensor_newWithSize(state, size, NULL);   \
-  THCudaTensor_copy##TYPEC(state, buffer, src);                         \
-  THCudaHalfTensor_copyCudaFloat(state, self, buffer);                  \
-  THCudaTensor_free(state, buffer);                                     \
-  THLongStorage_free(size);                                             \
-}
-#endif
 
 IMPLEMENT_TH_CUDA_TENSOR_COPY(Byte)
 IMPLEMENT_TH_CUDA_TENSOR_COPY(Char)
@@ -63,10 +47,10 @@ IMPLEMENT_TH_CUDA_TENSOR_COPY(Int)
 IMPLEMENT_TH_CUDA_TENSOR_COPY(Long)
 IMPLEMENT_TH_CUDA_TENSOR_COPY(Float)
 IMPLEMENT_TH_CUDA_TENSOR_COPY(Double)
+IMPLEMENT_TH_CUDA_TENSOR_COPY(Half)
 
 /* copyCuda */
 
-#ifndef THC_REAL_IS_HALF
 void THTensor_(copyCuda)(THCState *state, THTensor *self, struct THCTensor *src)
 {
   THArgCheck(THTensor_(nElement)(self) == THCTensor_(nElement)(state, src), 2, "sizes do not match");
@@ -84,9 +68,7 @@ void THTensor_(copyCuda)(THCState *state, THTensor *self, struct THCTensor *src)
     THTensor_(freeCopyTo)(selfc, self);
   }
 }
-#endif
 
-#ifndef THC_REAL_IS_HALF
 #define IMPLEMENT_TH_CUDA_TENSOR_COPY_TO(TYPEC)                           \
   void TH_CONCAT_4(TH,TYPEC,Tensor_copyCuda,Real)(THCState *state, TH##TYPEC##Tensor *self, struct THCTensor *src) \
   {                                                                       \
@@ -104,19 +86,6 @@ void THTensor_(copyCuda)(THCState *state, THTensor *self, struct THCTensor *src)
       THTensor_(free)(srcf);                                              \
     }                                                                     \
   }
-#else
-#define IMPLEMENT_TH_CUDA_TENSOR_COPY_TO(TYPEC)                           \
-  void TH_CONCAT_4(TH,TYPEC,Tensor_copyCuda,Real)(THCState *state, TH##TYPEC##Tensor *self, struct THCTensor *src) \
-  {                                                                       \
-    THArgCheck(TH##TYPEC##Tensor_nElement(self) == THCTensor_(nElement)(state, src), 2, "sizes do not match");     \
-    THLongStorage *size = THCTensor_(newSizeOf)(state, src);              \
-    THCudaTensor *buffer = THCudaTensor_newWithSize(state, size, NULL);   \
-    THCudaTensor_copyCudaHalf(state, buffer, src);                        \
-    TH_CONCAT_3(TH,TYPEC,Tensor_copyCudaFloat)(state, self, buffer);      \
-    THCudaTensor_free(state, buffer);                                     \
-    THLongStorage_free(size);                                             \
-  }
-#endif
 
 IMPLEMENT_TH_CUDA_TENSOR_COPY_TO(Byte)
 IMPLEMENT_TH_CUDA_TENSOR_COPY_TO(Char)
@@ -125,13 +94,13 @@ IMPLEMENT_TH_CUDA_TENSOR_COPY_TO(Int)
 IMPLEMENT_TH_CUDA_TENSOR_COPY_TO(Long)
 IMPLEMENT_TH_CUDA_TENSOR_COPY_TO(Float)
 IMPLEMENT_TH_CUDA_TENSOR_COPY_TO(Double)
+IMPLEMENT_TH_CUDA_TENSOR_COPY_TO(Half)
 
 void THCTensor_(copyCuda)(THCState *state, THCTensor *self, THCTensor *src)
 {
   THCTensor_(copy)(state, self, src);
 }
 
-#ifndef THC_REAL_IS_HALF
 void THCTensor_(copyAsyncCPU)(THCState *state, THCTensor *self, struct THTensor *src)
 {
   THArgCheck(THCTensor_(nElement)(state, self) == THTensor_(nElement)(src), 2, "sizes do not match");
@@ -193,7 +162,6 @@ void THTensor_(copyAsyncCuda)(THCState *state, THTensor *self, struct THCTensor 
     THCudaCheck(cudaSetDevice(currentDevice));
   }
 }
-#endif
 
 #undef IMPLEMENT_TH_CUDA_TENSOR_COPY
 #undef IMPLEMENT_TH_CUDA_TENSOR_COPY_TO
