+++ lib/THC/generic/THCTensorCopy.h
@@ -11,6 +11,7 @@ THC_API void THCTensor_(copyInt)(THCState *state, THCTensor *self, THIntTensor *
 THC_API void THCTensor_(copyLong)(THCState *state, THCTensor *self, THLongTensor *src);
 THC_API void THCTensor_(copyFloat)(THCState *state, THCTensor *self, THFloatTensor *src);
 THC_API void THCTensor_(copyDouble)(THCState *state, THCTensor *self, THDoubleTensor *src);
+THC_API void THCTensor_(copyHalf)(THCState *state, THCTensor *self, struct THHalfTensor *src);
 
 THC_API void THCTensor_(copyCudaByte)(THCState *state, THCTensor *dst, struct THCudaByteTensor *src);
 THC_API void THCTensor_(copyCudaChar)(THCState *state, THCTensor *dst, struct THCudaCharTensor *src);
@@ -30,15 +31,13 @@ THC_API void TH_CONCAT_2(THIntTensor_copyCuda   , Real)  (THCState *state, THInt
 THC_API void TH_CONCAT_2(THLongTensor_copyCuda  , Real)  (THCState *state, THLongTensor *self, THCTensor *src);
 THC_API void TH_CONCAT_2(THFloatTensor_copyCuda , Real)  (THCState *state, THFloatTensor *self, THCTensor *src);
 THC_API void TH_CONCAT_2(THDoubleTensor_copyCuda, Real)  (THCState *state, THDoubleTensor *self, THCTensor *src);
+THC_API void TH_CONCAT_2(THHalfTensor_copyCuda, Real)    (THCState *state, THHalfTensor *self, THCTensor *src);
 THC_API void THCTensor_(copyCuda) (THCState *state, THCTensor *self, THCTensor *src);
 
-/* There is no THHalfTensor */
-#ifndef THC_REAL_IS_HALF
 THC_API void THTensor_(copyCuda) (THCState *state, THTensor *self, THCTensor *src);
 THC_API void THCTensor_(copyCPU) (THCState *state, THCTensor *self, THTensor *src);
 
 THC_API void THCTensor_(copyAsyncCPU)(THCState *state, THCTensor *self, THTensor *src);
 THC_API void THTensor_(copyAsyncCuda)(THCState *state, THTensor *self, THCTensor *src);
-#endif
 
 #endif
