+++ torch/generic/Tensor.c
@@ -716,6 +716,8 @@ static int torch_Tensor_(copy)(lua_State *L)
     THCTensor_(copyFloat)(state, tensor, src);
   else if( (src = luaT_toudata(L, 2, "torch.DoubleTensor")) )
     THCTensor_(copyDouble)(state, tensor, src);
+  else if( (src = luaT_toudata(L, 2, "torch.HalfTensor")) )
+    THCTensor_(copyHalf)(state, tensor, src);
   else
     luaL_typerror(L, 2, "torch.*Tensor");
   lua_settop(L, 1);
@@ -793,6 +795,11 @@ static int torch_Tensor_(__newindex__)(lua_State *L)
       tensor = THCTensor_(newWithTensor)(state, tensor);
       THCTensor_(narrow)(state, tensor, NULL, 0, index, 1);
       THCTensor_(copyDouble)(state, tensor, src);
+      THCTensor_(free)(state, tensor);
+    } else if( (src = luaT_toudata(L, 3, "torch.HalfTensor")) ) {
+      tensor = THCTensor_(newWithTensor)(state, tensor);
+      THCTensor_(narrow)(state, tensor, NULL, 0, index, 1);
+      THCTensor_(copyHalf)(state, tensor, src);
       THCTensor_(free)(state, tensor);
     } else {
       luaL_typerror(L, 3, "torch.*Tensor");
@@ -913,6 +920,8 @@ static int torch_Tensor_(__newindex__)(lua_State *L)
         THCTensor_(copyFloat)(state, tensor, src);
       } else if( (src = luaT_toudata(L, 3, "torch.DoubleTensor")) ) {
         THCTensor_(copyDouble)(state, tensor, src);
+      } else if( (src = luaT_toudata(L, 3, "torch.HalfTensor")) ) {
+        THCTensor_(copyHalf)(state, tensor, src);
       } else {
         luaL_typerror(L, 3, "torch.*Tensor");
       }
